refactor(target): target_amount decimal(15,0) + commission_rate decimal:2
Match the project-wide money column convention (decimal) and tighten the
commission rate cast for precision.

+++ database/migrations/2026_06_22_000001_create_tb_marketing_targets_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->smallInteger('year');
             $table->unsignedTinyInteger('month');
-            $table->unsignedBigInteger('target_amount')->default(0);
+            $table->decimal('target_amount', 15, 0)->default(0);
             $table->decimal('commission_rate', 5, 2)->default(0);
             $table->foreignId('created_by')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('updated_by')->nullable()->constrained('users')->nullOnDelete();
